# cmd/swag/docs/swagger.yaml
basePath: /v1/2025
definitions:
...
  model.RoleAccount:
    properties:
      account_id:
        example: b16fdfbc-da22-4ab2-b045-614677e536d6
        type: string
      license_id:
        example: xxx-yyyy-zzz
        type: string
      role_id:
        example: 351ec27d-2134-4993-b120-b72a950f5574
        type: string
    required:
    - account_id
    - role_id
    type: object
  model.RolesMenu:
    properties:
      list_method:
        example:
        - '[''GET'''
        - '''POST'']'
        items:
          type: string
        type: array
      menu_id:
        example: menu-123
        type: string
      role_id:
        example: role-123
        type: string
    required:
    - list_method
    - menu_id
    - role_id
    type: object
...
  response.ErrorResponseData:
    properties:
      code:
        description: status code
        type: integer
      detail:
        description: dữ liệu được return
      error:
        description: Thông báo lỗi
        type: string
    type: object
  response.ResponseData:
    properties:
      code:
        description: status code
        type: integer
      data:
        description: dữ liệu được return
      message:
        description: Thông báo lỗi
        type: string
    type: object
host: localhost:8002
info:
  contact:
    email: teamaithacoindustries@gmail.com
    name: TEAM AI
    url: github.com/phananhtu1998/go-backend-api
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: github.com/phananhtu1998/go-backend-api
  title: API DOCUMENTATION PARKING BACKEND
  version: 1.0.0
paths:
...
  /roleaccount/create_roles_account:
    post:
      consumes:
      - application/json
      description: Api tạo role account cho hệ thống
      parameters:
      - description: Role account mapping details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RoleAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Tạo role account
      tags:
      - RoleAccount
  /roleaccount/delete_multiple_role_account:
    delete:
      consumes:
      - application/json
      description: API này xóa danh sách Role Account dựa trên danh sách ID
      parameters:
      - description: Danh sách ID cần xóa
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Xóa nhiều Role Account
      tags:
      - RoleAccount
  /roleaccount/get_role_account_by_account_id/{id}:
    get:
      consumes:
      - application/json
      description: API này trả về role account theo Account_Id
      parameters:
      - description: ID account
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Lấy role account theo Account_Id
      tags:
      - RoleAccount
  /roleaccount/get_role_account_by_role_id/{id}:
    get:
      consumes:
      - application/json
      description: API này trả về role account theo role_Id
      parameters:
      - description: ID role
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Lấy role account theo Role_Id
      tags:
      - RoleAccount
  /roleaccount/update_role_account/{id}:
    put:
      consumes:
      - application/json
      description: API này cập nhật thông tin role account dựa trên ID
      parameters:
      - description: ID role account cần cập nhật
        in: path
        name: id
        required: true
        type: string
      - description: Dữ liệu cập nhật role account
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RoleAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Cập nhật Role account
      tags:
      - RoleAccount
  /rolesmenu/create_roles_menu:
    post:
      consumes:
      - application/json
      description: Api tạo role menu cho hệ thống
      parameters:
      - description: Role menu mapping details
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RolesMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Tạo role menu
      tags:
      - RolesMenu
  /rolesmenu/delete_roles_menu/{id}:
    delete:
      consumes:
      - application/json
      description: Api xóa role menu
      parameters:
      - description: ID của role menu
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Xóa role menu
      tags:
      - RolesMenu
  /rolesmenu/get_role_menu_by_role_id/{id}:
    get:
      consumes:
      - application/json
      description: API này trả về role menu theo ID
      parameters:
      - description: ID role
        in: path
        name: id
        required: true
        type: string
      - description: Từ khóa tìm kiếm
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Lấy role menu theo ID
      tags:
      - RolesMenu
  /rolesmenu/update_roles_menu/{id}:
    put:
      consumes:
      - application/json
      description: Api cập nhật phân quyền menu cho role
      parameters:
      - description: ID của role menu
        in: path
        name: id
        required: true
        type: string
      - description: Thông tin cần cập nhật
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.RolesMenu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponseData'
      security:
      - BearerAuth: []
      - ApiKeyAuth: []
      summary: Cập nhật role menu
      tags:
      - RolesMenu
...
security:
- "":
  - BearerAuth
  - ApiKeyAuth
securityDefinitions:
  ApiKeyAuth:
    description: Nhập API Key vào header để xác thực API
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Nhập Bearer Token vào header. Ví dụ: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"